Fix all golangci-lint errors and make coverage badge non-blocking
Fix lint issues across 58 files: unnecessary type conversions, type
switch simplifications, misspellings, gofmt/goimports formatting,
unused code removal, raw string regexp, context key types, and
redundant returns. Add continue-on-error to coverage badge step
(requires wiki to be initialized). Exclude generated PEG parser
from lint.

+++ aws/fetch/manual_fetchers.go
@@ -784,7 +784,7 @@ func addManualMessagingFetchFuncs(conf *Config, funcs map[string]fetch.Func) {
 							if err != nil {
 								errC <- err
 							}
-							res.Properties()[properties.Created] = time.Unix(int64(timestamp), 0)
+							res.Properties()[properties.Created] = time.Unix(timestamp, 0)
 						}
 					case "LastModifiedTimestamp":
 						if v != "" {
@@ -792,7 +792,7 @@ func addManualMessagingFetchFuncs(conf *Config, funcs map[string]fetch.Func) {
 							if err != nil {
 								errC <- err
 							}
-							res.Properties()[properties.Modified] = time.Unix(int64(timestamp), 0)
+							res.Properties()[properties.Modified] = time.Unix(timestamp, 0)
 						}
 					case "QueueArn":
 						res.Properties()[properties.Arn] = v
